Added JS implementation of AddProjectedIncome and AddBudget
Added Javascript functions for initial user set-up, adding the AddProjectedIncome method to input their projected income, the AddBudget method to input their budget, and revised the AddBills method to alert the user to the contents of a bill after they have inputted the data. Buttons in dashboard.html have also been added to call this functions on user click.